refactor(api): serve resolve and check under /components
/external named the implementation (a call to an external provider)
rather than the operand. Both endpoints operate on components — one
resolves a FetchField's options from candidate credentials, one checks
a connection's candidate config — so they move to POST
/components/resolve and POST /components/check, and the routes/components
module becomes a package (crud + resolve + check) owning every component
operation. A future saved-connection check lands beside them as
POST /components/{id}/check.

By Digitl
