+++ packages/interloper-api/tests/test_components_check.py
@@ -1,4 +1,4 @@
-"""Tests for ``interloper_api.routes.external.check`` (generic connection checker)."""
+"""Tests for ``interloper_api.routes.components.check`` (generic connection checker)."""
 
 from __future__ import annotations
 
@@ -12,7 +12,7 @@ from interloper_assets.facebook_ads.connection import FacebookAdsConnection
 from interloper_assets.facebook_ads.source import FacebookAds
 
 from interloper_api.dependencies import get_catalog, require_viewer
-from interloper_api.routes import external as external_module
+from interloper_api.routes import components as components_module
 
 CONFIG = {"access_token": "TOK", "app_id": "A", "app_secret": "S", "_id": "x"}
 
@@ -25,7 +25,7 @@ class UncheckableConnection(il.Connection):
 
 def _client(catalog: il.Catalog) -> TestClient:
     app = FastAPI()
-    app.include_router(external_module.router, prefix="/external")
+    app.include_router(components_module.router, prefix="/components")
     app.dependency_overrides[require_viewer] = lambda: None
     app.dependency_overrides[get_catalog] = lambda: catalog
     return TestClient(app)
@@ -54,7 +54,7 @@ def mock_graph(monkeypatch: pytest.MonkeyPatch):
 
 def _check(catalog: il.Catalog, config: dict) -> httpx.Response:
     return _client(catalog).post(
-        "/external/check",
+        "/components/check",
         json={"component_key": "facebook_ads_connection", "config": config},
     )
 
@@ -96,7 +96,7 @@ class TestCheck:
     def test_uncheckable_connection_is_static_only(self):
         catalog = il.Catalog(components={UncheckableConnection.key: UncheckableConnection.definition()})
         resp = _client(catalog).post(
-            "/external/check",
+            "/components/check",
             json={"component_key": "uncheckable_connection", "config": {"api_key": "k"}},
         )
 
@@ -104,10 +104,10 @@ class TestCheck:
         assert (body["ok"], body["live"]) == (True, False)
 
     def test_unknown_component_404(self, catalog: il.Catalog):
-        resp = _client(catalog).post("/external/check", json={"component_key": "nope", "config": {}})
+        resp = _client(catalog).post("/components/check", json={"component_key": "nope", "config": {}})
         assert resp.status_code == 404
 
     def test_non_connection_component_404(self, catalog: il.Catalog):
         catalog = il.Catalog.from_assets([FacebookAds])
-        resp = _client(catalog).post("/external/check", json={"component_key": "facebook_ads", "config": {}})
+        resp = _client(catalog).post("/components/check", json={"component_key": "facebook_ads", "config": {}})
         assert resp.status_code == 404
